Answers for part 6

+++ app/controllers/questions_controller.rb
@@ -1,19 +1,24 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [ :index, :show ]
+  before_action :set_question, only: [:show, :edit, :update, :destroy]
+
   def index
     @questions = Question.all
   end
 
-  def show; end
+  def show 
+    @answer = Answer.new  
+  end
 
   def new; end
 
   def edit; end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: "Your question was succesfully created"
     else
       render :new
     end
@@ -28,11 +33,19 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to question_path
+    if current_user.author?(question)
+      question.destroy
+      redirect_to question_path, notice: "Question successfully deleted."
+    else
+      redirect_to question, notice: "The question has not been deleted. You are not the author of the question."
+    end
   end
 
   private
+
+  def set_question
+    @question = Question.find(params[:id])
+  end
 
   def question
     @question ||= params[:id] ? Question.find(params[:id]) : Question.new
